requirements.txt
# Scraping
google-play-scraper
requests

# Data Processing
pandas
numpy

# Visualization (notebooks + console)
matplotlib
seaborn
plotly

# Console UI
rich

# LLM / AI
google-generativeai
openai>=1.0.0  # OpenRouter uses OpenAI-compatible API

# Email
resend

# Database
supabase

# Utils
python-dotenv
